refactor(notifications): extract billing-limit predicate into lib/billing helper (KEEP-448)
Move isBillingLimitReached out of the API route and into lib/billing/limit-status.ts
so the SQL and plan-resolution logic isn't living in the controller. Route now just
orchestrates auth + delegation.
