+++ app/api/notifications/status/route.ts
@@ -1,14 +1,6 @@
-import { and, count, eq, sql } from "drizzle-orm";
 import { NextResponse } from "next/server";
 import { isBillingEnabled } from "@/lib/billing/feature-flag";
-import {
-  getPlanLimits,
-  parsePlanName,
-  parseTierKey,
-} from "@/lib/billing/plans";
-import { getOrgSubscription } from "@/lib/billing/plans-server";
-import { db } from "@/lib/db";
-import { workflows } from "@/lib/db/schema";
+import { isBillingLimitReached } from "@/lib/billing/limit-status";
 import { ErrorCategory, logSystemError } from "@/lib/logging";
 import {
   auditFromAuth,
@@ -39,11 +31,8 @@ export async function GET(
 
     const types: NotificationType[] = [];
 
-    if (isBillingEnabled()) {
-      const billingLimitReached = await isBillingLimitReached(organizationId);
-      if (billingLimitReached) {
-        types.push("billing_limit_reached");
-      }
+    if (isBillingEnabled() && (await isBillingLimitReached(organizationId))) {
+      types.push("billing_limit_reached");
     }
 
     return NextResponse.json({ unreadCount: types.length, types });
@@ -63,55 +52,4 @@ export async function GET(
       { status: 500 }
     );
   }
-}
-
-async function isBillingLimitReached(organizationId: string): Promise<boolean> {
-  const sub = await getOrgSubscription(organizationId);
-  const plan = parsePlanName(sub?.plan);
-  const tier = parseTierKey(sub?.tier);
-  const limits = getPlanLimits(plan, tier, sub?.planOverrides);
-
-  if (limits.maxExecutionsPerMonth < 0) {
-    return false;
-  }
-
-  const enabledWorkflowsResult = await db
-    .select({ value: count() })
-    .from(workflows)
-    .where(
-      and(
-        eq(workflows.organizationId, organizationId),
-        eq(workflows.enabled, true)
-      )
-    );
-  const enabledWorkflows = enabledWorkflowsResult[0]?.value ?? 0;
-  if (enabledWorkflows === 0) {
-    return false;
-  }
-
-  const startOfMonth = new Date();
-  startOfMonth.setUTCDate(1);
-  startOfMonth.setUTCHours(0, 0, 0, 0);
-
-  const usageResult = await db.execute<{ count: number }>(
-    sql`SELECT
-          (
-            SELECT COUNT(*)
-              FROM workflow_executions we
-              JOIN workflows w ON we.workflow_id = w.id
-             WHERE w.organization_id = ${organizationId}
-               AND we.started_at >= ${startOfMonth.toISOString()}
-               AND we.status <> 'blocked_billing'
-          )
-          +
-          (
-            SELECT COUNT(*)
-              FROM direct_executions de
-             WHERE de.organization_id = ${organizationId}
-               AND de.created_at >= ${startOfMonth.toISOString()}
-               AND de.status <> 'blocked_billing'
-          ) AS count`
-  );
-  const used = usageResult[0]?.count ?? 0;
-  return used >= limits.maxExecutionsPerMonth;
 }
